refactor: Update dumpJson.ps1 to include customizable depth parameter
This commit updates the dumpJson.ps1 file to include a customizable depth parameter. The depth parameter allows users to specify the level of nested objects to include when converting to JSON. By default, the depth is set to 10, but now users can provide a different value. This enhancement improves the flexibility and usability of the dumpJson function in the PowerShell module.