feat: Add 'Compress' option to dumpJson.ps1
This commit adds a new 'Compress' option to the dumpJson.ps1 script, allowing the user to compress the JSON output. The 'Compress' option is a switch parameter that can be used to reduce the size of the JSON output by removing unnecessary whitespace. This can be useful when dealing with large JSON objects.